Revert some changes from r2206.
That revision is partially a wrong merge. The pt-BR translation team
decided to fix the translation at the version 1.2 of the book. However,
that merge included changes related to the version 1.3. Only those changes
were reverted, namely the ones from r2160. Changes from r2142 needed not
to be reverted because they were not included in the merge.

git-svn-id: https://svn.code.sf.net/p/svnbook/source/trunk@2279 b70f5e92-ccc6-4167-9ab2-d027528d294b